% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_step_ahead_model_output.R
\name{output_plot}
\alias{output_plot}
\title{Plot projection data}
\usage{
output_plot(
  all_plot,
  all_ens,
  truth_data,
  plot_truth = TRUE,
  intervals = c(0.5, 0.8, 0.95),
  pal_color = "Set2",
  fill_transparency = 0.25,
  pal_value = NULL,
  top_layer = "model_output",
  ens_color = NULL,
  ens_name = NULL,
  facet = NULL,
  facet_scales = "fixed",
  facet_nrow = NULL,
  facet_ncol = NULL,
  facet_title = "top left",
  facet_value = NULL,
  interactive = TRUE,
  fill_by = "model_id",
  x_col_name = "target_date",
  x_truth_col_name = "time_idx"
)
}
\arguments{
\item{all_plot}{a list with two data frame: one for plain lines,
one for ribbons plotting (in a wide format)}

\item{all_ens}{a list with two data frame: one for plain lines,
one for ribbons plotting (in a wide format) for a unique \code{model_id} value
associated with specific color (\code{ens_color}). NULL is no specific layout
required}

\item{truth_data}{a \code{data.frame} object containing the ground truth data,
containing the columns: date information (\code{x_truth_col_name} parameter) and
\code{value}. Ignored, if \code{plot_truth = FALSE}.}

\item{plot_truth}{a \code{boolean} for showing the truth data in the plot.
Default to TRUE. Data used in the plot comes from the parameter \code{truth_data}}

\item{intervals}{a vector of \code{numeric} values indicating which central
prediction interval levels to plot. \code{NULL} means no interval levels.
If not provided, it will default to \code{c(.5, .8, .95)}.
When plotting 6 models or more, the plot will be reduced to show \code{.95}
interval only. Value possibles: \verb{0.5, 0.8, 0.9, 0.95}}

\item{pal_color}{a \code{character} string for specifying the palette color in the
plot. For \code{plotly} plots, please refer to
\code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::display.brewer.all()}}. Default to \code{"Set2"}}

\item{fill_transparency}{numeric value used to set transparency of intervals.
0 means fully transparent, 1 means opaque. Default to \code{0.25}}

\item{pal_value}{a \verb{named vector} containing the \code{model_id} (names) and
associated color. Default \code{NULL}, used only for static plot (ggplot2)}

\item{top_layer}{character vector, where the first element indicates the top
layer of the resulting plot. Possible options are \code{"model_output"} (default)
and \code{"truth"}}

\item{ens_color}{a \code{character} string of a color name, if not NULL, will be
use as color for the model name associated with the parameter \code{ens_name}}

\item{ens_name}{a \code{character} string of a model name, if not NULL, will be
use to change the color for the model name, associated with the parameter
\code{ens_color}(both parameter need to be provided)}

\item{facet}{a unique value corresponding as a task_id variable name
(interpretable as facet option for ggplot)}

\item{facet_scales}{argument for scales as in \link[ggplot2:facet_wrap]{ggplot2::facet_wrap} or
equivalent to \code{shareX}, \code{shareY} in \link[plotly:subplot]{plotly::subplot}. Default to "fixed"
(x and y axes are shared).}

\item{facet_nrow}{a numeric, number of rows in the layout.}

\item{facet_ncol}{a numeric, number of columns in the layout.}

\item{facet_title}{a \code{string}, position of each subplot tile (value
associated with the \code{facet} parameter). "top right", "top left" (default),
"bottom right", "bottom left" are the possible values, \code{NULL} to remove the
title. For interactive plot only.}

\item{facet_value}{a vector of all the possible unique values in the
associated column \code{facet}}

\item{interactive}{a \code{boolean} to output an "interactive" version of the
plot (using Plotly) or a "static" plot (using ggplot2). By default, \code{TRUE}
(interactive plot)}

\item{fill_by}{name of a column for specifying colors and legend in plot.
The \code{pal_color} parameter can be use to change the palette.
Default to \code{model_id}.}

\item{x_col_name}{column name containing the date information for \code{all_plot}
and \code{all_ens} data frames, value will be map to the x-axis of the plot.
By default, "target_date".}

\item{x_truth_col_name}{column name containing the date information for
\code{truth_data} data frame, value will be map to the x-axis of the plot.
By default, "time_idx".}
}
\description{
Use Plotly or ggplot2 to plot projection model output with or without
truth data.
}
