% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_step_ahead_forecasts.R
\name{plot_step_ahead_forecasts}
\alias{plot_step_ahead_forecasts}
\title{Basic Plot for model outputs}
\usage{
plot_step_ahead_forecasts(
  forecast_data,
  truth_data,
  plain_line = 0.5,
  ribbon = c(0.975, 0.025),
  title = NULL,
  ens_color = NULL,
  ens_name = NULL
)
}
\arguments{
\item{forecast_data}{a \code{data.frame} object containing the columns: \code{model_id},
\code{output_type_id}, \code{target_date}, \code{value}}

\item{truth_data}{a \code{data.frame} object containing the ground truth data,
containing the columns: \code{time_idx} and \code{value}}

\item{plain_line}{a \code{numeric} output_type_id value, value will be used to
create a plain line in the plot. Should be a unique value}

\item{ribbon}{a vector of \code{numeric}  output_type_id value, value will be
used a to create a ribbon in the plot.}

\item{title}{a \code{character} string, if not NULL, will be added as title to the
plot}

\item{ens_color}{a \code{character} string of a color name, if not NULL, will be
use as color for the model name associated with the parameter \code{ens_name}
(both parameter need to be provided)}

\item{ens_name}{a \code{character} string of a model name, if not NULL, will be
use to change the color for the model name, associated with the parameter
\code{ens_color}(both parameter need to be provided)}
}
\description{
Create a simple Plotly time-series plot for model projection outputs.
}
