% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_step_ahead_forecasts.R
\name{plotly_model_plot}
\alias{plotly_model_plot}
\title{Plot forecast data with Plotly}
\usage{
plotly_model_plot(
  plot_model,
  df_point,
  df_ribbon,
  plot_truth,
  truth_data,
  opacity = 0.25,
  line_color = NULL,
  ...
)
}
\arguments{
\item{plot_model}{a plot_ly object to add lines and/or ribbons, if NULL will
create an empty object}

\item{df_point}{a \code{data.frame} with "target_date" and "value" columns, use to
add lines on the plot}

\item{df_ribbon}{a \code{data.frame} with "target_date", "min", and "max" columns,
use to add ribbons on the plot}

\item{plot_truth}{a \code{boolean} for showing the truth data in the plot.
Default to TRUE. Data used in the plot comes from the parameter \code{truth_data}}

\item{truth_data}{a \code{data.frame} object containing the ground truth data,
containing the columns: \code{time_idx} and \code{value}.
Ignored, if \code{plot_truth = FALSE}}

\item{opacity}{a \code{numeric}, opacity of the ribbons, default 0.25}

\item{line_color}{a \code{string}, specific color associated with plot}

\item{...}{ploty parameters}
}
\description{
Use Plotly to plot projection model output
}
