% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_step_ahead_forecasts.R
\name{plotly_plot}
\alias{plotly_plot}
\title{Plot  projection data with Plotly}
\usage{
plotly_plot(
  all_plot,
  all_ens,
  plot_truth,
  truth_data,
  intervals,
  pal_color,
  fill_transparency,
  top_layer,
  ens_color,
  facet,
  facet_scales,
  facet_nrow,
  facet_title,
  facet_value
)
}
\arguments{
\item{all_plot}{a list with two data frame: one for plain lines,
one for ribbons plotting (in a wide format)}

\item{all_ens}{a list with two data frame: one for plain lines,
one for ribbons plotting (in a wide format) for a unique \code{model_id} value
associated with specific color (\code{ens_color}). NULL is no specific layout
required}

\item{plot_truth}{a \code{boolean} for showing the truth data in the plot.
Default to TRUE. Data used in the plot comes from the parameter \code{truth_data}}

\item{truth_data}{a \code{data.frame} object containing the ground truth data,
containing the columns: \code{time_idx} and \code{value}.
Ignored, if \code{plot_truth = FALSE}}

\item{intervals}{a vector of \code{numeric} values indicating which central
prediction interval levels to plot. \code{NULL} means no interval levels.
If not provided, it will default to \code{c(.5, .8, .95)}.
When plotting 6 models or more, the plot will be reduced to show \code{.95}
interval only. Value possibles: \verb{0.5, 0.8, 0.9, 0.95}}

\item{pal_color}{a \code{character} string for specifying the palette color in the
plot if \code{fill_by_model} is set to \code{TRUE}. For \code{plotly} plots, please refer
to \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::display.brewer.all()}}. Default to \code{"Set2"}}

\item{fill_transparency}{numeric value used to set transparency of intervals.
0 means fully transparent, 1 means opaque. Default to \code{0.25}}

\item{top_layer}{character vector, where the first element indicates the top
layer of the resulting plot. Possible options are \code{"forecast"} (default)
and \code{"truth"}}

\item{ens_color}{a \code{character} string of a color name, if not NULL, will be
use as color for the model name associated with the parameter \code{ens_name}}

\item{facet}{a unique value corresponding as a task_id variable name
(interpretable as facet option for ggplot)}

\item{facet_scales}{argument for scales as in \link[ggplot2:facet_wrap]{ggplot2::facet_wrap} or
equivalent to \code{shareX}, \code{shareY} in \link[plotly:subplot]{plotly::subplot}. Default to "fixed"
(x and y axes are shared).}

\item{facet_nrow}{a numeric, number of rows in the layout.}

\item{facet_title}{a \code{string}, position of each subplot tile (value
associated with the \code{facet} parameter). "top right", "top left" (default),
"bottom right", "bottom left" are the possible values, \code{NULL} to remove the
title}

\item{facet_value}{a vector of all the possible unique values in the
associated column \code{facet}}
}
\description{
Use Plotly to plot projection model output with or without truth data.
}
