% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_step_ahead_forecasts.R
\name{plotly_truth_data}
\alias{plotly_truth_data}
\title{Plot Truth data with Plotly}
\usage{
plotly_truth_data(plot_model, truth_data, plot_truth, arguments)
}
\arguments{
\item{plot_model}{a plot_ly object to add lines and/or ribbons, if NULL will
create an empty object}

\item{truth_data}{a \code{data.frame} object containing the ground truth data,
containing the columns: \code{time_idx} and \code{value}.
Ignored, if \code{plot_truth = FALSE}}

\item{plot_truth}{a \code{boolean} for showing the truth data in the plot.
Default to TRUE. Data used in the plot comes from the parameter \code{truth_data}}

\item{arguments}{list of others Plotly parameters}
}
\description{
Use Plotly to plot truth data
}
