% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scenario_data}
\alias{scenario_data}
\alias{scenario_outputs}
\alias{scenario_target_ts}
\title{Example scenario hub data}
\format{
\subsection{\code{scenario_outputs}}{

A data frame with 7,176 rows and 9 columns:
\describe{
\item{model_id}{the name of the model}
\item{origin_date}{the starting point of the projection in yyyy-mm-dd
format}
\item{scenario_id}{a unique identifier for the scenario}
\item{location}{FIPS code identifying a location}
\item{target}{a unique identifier for the target}
\item{horizon}{number of time units ahead being projected relative to the
\code{origin_date}, in units of weeks}
\item{output_type}{the type of representation of the prediction; in this
example, all values for the \code{output_type} are "quantile".}
\item{output_type_id}{more identifying information specific to the
output type; here, the \code{output_type_id} specifies the probability
level for the quantile prediction}
\item{value}{the model’s prediction}

}
}

\subsection{\code{scenario_target_ts}}{

A data frame with 127 rows and 3 columns:
\describe{
\item{location}{FIPS code identifying a location}
\item{date}{the date of the target observation in yyyy-mm-dd format}
\item{observation}{a count of incident cases in the given \code{location}
in the week ending on the given \code{date}.}
\item{target}{a unique identifier for the target}
}
}
}
\source{
\url{https://github.com/hubverse-org/example-complex-scenario-hub/}
}
\usage{
scenario_outputs

scenario_target_ts
}
\description{
The hubExamples package provides two data sets that contain example model output and
target data for an example scenario hub: \code{scenario_outputs} and \code{scenario_target_ts}.

Data from \code{hubExample} package: https://github.com/hubverse-org/hubExamples

\code{scenario_target_ts} contains time series target data associated with the
scenario projection data.
}
\details{
\code{scenario_outputs} contains example scenario projection data that represents
model outputs and an ensemble from a scenario hub with predictions for one
target (\verb{inc hosp}) in one location (\code{"US"}), one round ("2021-03-07") and
four scenarios. This dataset reflects scenario projection outputs as they look
when retrieved from a hub via the \code{hubData} package (which is slightly
different than they look when originally submitted by modelers), and with
a "mean" ensemble calculated with the \code{hubEnsemble} package
\code{simple_ensemble()} function with default parameters.
The date of occurrence of the outcome of interest in yyyy-mm-dd format
can be calculated directly from the \code{origin_date} and \code{horizon} as follows:
\code{target_end_date = origin_date + (7 * horizon) - 1}

Data from \code{hubExample} package: https://github.com/hubverse-org/hubExamples
}
\keyword{datasets}
