% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_step_ahead_model_output.R
\name{simple_model_plot}
\alias{simple_model_plot}
\title{Plot simple projection data}
\usage{
simple_model_plot(
  plot_model,
  df_point,
  df_ribbon,
  plot_truth,
  truth_data,
  opacity = 0.25,
  line_color = NULL,
  top_layer = "model_output",
  show_truth_legend = TRUE,
  interactive = TRUE,
  fill_by = "model_id",
  x_col_name = "target_date",
  x_truth_col_name = "time_idx",
  ...
)
}
\arguments{
\item{plot_model}{a plot_ly object to add lines and/or ribbons, if NULL will
create an empty object.}

\item{df_point}{a \code{data.frame} with a column containing date information
(\code{x_col_name} parameter) and "value" columns, use to add lines on the plot.}

\item{df_ribbon}{a \code{data.frame} with a column containing date information
(\code{x_col_name} parameter), and a "min", and "max" columns, use to add ribbons
on the plot.}

\item{plot_truth}{a \code{boolean} for showing the truth data in the plot.
Default to TRUE. Data used in the plot comes from the parameter \code{truth_data}}

\item{truth_data}{a \code{data.frame} object containing the ground truth data,
containing the columns: date information (\code{x_truth_col_name} parameter) and
\code{value}. Ignored, if \code{plot_truth = FALSE}.}

\item{opacity}{a \code{numeric}, opacity of the ribbons, default 0.25.}

\item{line_color}{a \code{string}, specific color associated with plot.}

\item{top_layer}{character vector, where the first element indicates the top
layer of the resulting plot. Possible options are \code{"model_output"} (default)
and \code{"truth"}.}

\item{show_truth_legend}{a \code{boolean} to show legend of the truth data, by
default \code{TRUE}.}

\item{interactive}{a \code{boolean} to output an "interactive" version of the
plot (using Plotly) or a "static" plot (using ggplot2). By default, \code{TRUE}
(interactive plot).}

\item{fill_by}{name of a column for specifying colors and legend in plot.
The \code{pal_color} parameter can be use to change the palette.
Default to \code{model_id}.}

\item{x_col_name}{column name containing the date information for \code{df_point}
and \code{df_ribbon} data frames, value will be map to the x-axis of the plot.
By default, "target_date".}

\item{x_truth_col_name}{column name containing the date information for
\code{truth_data} data frame, value will be map to the x-axis of the plot.
By default, "time_idx".}

\item{...}{additional Plotly parameters.}
}
\description{
Use Plotly or ggplot2 to plot simple projection model output with or
without truth data.
Simple projection model output are defined as projection associated with one
particular set of "tasks_ids" value. For more information, please refer to
\href{https://hubdocs.readthedocs.io/en/latest/format/tasks.html}{HubDocs website}.
}
