% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_step_ahead_model_output.R
\name{static_proj_data}
\alias{static_proj_data}
\title{Plot Projection data with GGPLOT2}
\usage{
static_proj_data(plot_model, df_point, df_ribbon, line_color, opacity)
}
\arguments{
\item{plot_model}{a plot_ly object to add lines and/or ribbons, if NULL will
create an empty object}

\item{df_point}{a \code{data.frame} with "target_date" and "value" columns, use to
add lines on the plot}

\item{df_ribbon}{a \code{data.frame} with "target_date", "min", and "max" columns,
use to add ribbons on the plot}

\item{line_color}{a \code{string}, specific color associated with plot}

\item{opacity}{a \code{numeric}, opacity of the ribbons, default 0.25}
}
\description{
Use ggplot2 to plot projection model output
}
