% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_step_ahead_forecasts.R
\name{plot_prep_data}
\alias{plot_prep_data}
\title{Data Process for plotting function}
\usage{
plot_prep_data(df, plain_line, plain_type, intervals)
}
\arguments{
\item{df}{a \code{data.frame} object containing the columns: \code{model_id},
\code{output_type_id}, \code{target_date}, \code{value}}

\item{plain_line}{a \code{numeric} output_type_id value, value will be used to
create a plain line in the plot. Should be a unique value
(for example: 0.5)}

\item{plain_type}{a \code{string} output_type value, value will be used to
create a plain line in the plot. Should be a unique value
(for example: "quantile")}

\item{intervals}{a named list of \code{numeric}  output_type_id value, value will
be used a to create one or multiple prediction intervals in the plot.}
}
\description{
Data process for plotting function, returns a list with two
data frame: one for plain lines, one for ribbons plotting (in a wide format)
}
