% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_step_ahead_forecasts.R
\name{plot_step_ahead_forecasts}
\alias{plot_step_ahead_forecasts}
\title{Basic Plot for model outputs}
\usage{
plot_step_ahead_forecasts(
  forecast_data,
  truth_data,
  use_median_as_point = FALSE,
  plot = TRUE,
  plot_truth = TRUE,
  show_legend = TRUE,
  facet = NULL,
  facet_scales = "fixed",
  facet_nrow = NULL,
  facet_ncol = NULL,
  facet_title = "top left",
  interactive = TRUE,
  fill_by_model = TRUE,
  pal_color = "Set2",
  fill_transparency = 0.25,
  intervals = c(0.5, 0.8, 0.95),
  top_layer = "forecast",
  title = NULL,
  ens_color = NULL,
  ens_name = NULL
)
}
\arguments{
\item{forecast_data}{a \code{model_output_df} object, containing all the required
columns, and a "target_date" and a "model_id" column.}

\item{truth_data}{a \code{data.frame} object containing the ground truth data,
containing the columns: \code{time_idx} and \code{value}.
Ignored, if \code{plot_truth = FALSE}}

\item{use_median_as_point}{a \code{Boolean} for using median quantile as point
forecasts in plot. Default to FALSE. If TRUE, will select first any \code{median}
output type value and if no \code{median} value included in \code{forecast_data}; will
select \code{quantile = 0.5} output type value.}

\item{plot}{a \code{boolean} for showing the plot. Default to TRUE.}

\item{plot_truth}{a \code{boolean} for showing the truth data in the plot.
Default to TRUE. Data used in the plot comes from the parameter \code{truth_data}}

\item{show_legend}{a \code{boolean} for showing the legend in the plot.
Default to TRUE.}

\item{facet}{a unique value corresponding as a task_id variable name
(interpretable as facet option for ggplot)}

\item{facet_scales}{argument for scales as in \link[ggplot2:facet_wrap]{ggplot2::facet_wrap} or
equivalent to \code{shareX}, \code{shareY} in \link[plotly:subplot]{plotly::subplot}. Default to "fixed"
(x and y axes are shared).}

\item{facet_nrow}{a numeric, number of rows in the layout.}

\item{facet_ncol}{a numeric, number of columns in the layout
(ignored in \link[plotly:subplot]{plotly::subplot})}

\item{facet_title}{a \code{string}, position of each subplot tile (value
associated with the \code{facet} parameter). "top right", "top left" (default),
"bottom right", "bottom left" are the possible values, \code{NULL} to remove the
title. For interactive plot only.}

\item{interactive}{a \code{boolean} to output an "interactive" version of the
plot (using Plotly) or a "static" plot (using ggplot2). By default, \code{TRUE}
(interactive plot)}

\item{fill_by_model}{a \code{boolean} for specifying colors in plot. If \code{TRUE},
separate colors will be used for each model, \code{pal_color} parameters to change
the palette. If \code{FALSE}, only blues will be used for all models.
Default to \code{FALSE}.}

\item{pal_color}{a \code{character} string for specifying the palette color in the
plot if \code{fill_by_model} is set to \code{TRUE}. Please refer
to \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::display.brewer.all()}}. Default to \code{"Set2"}}

\item{fill_transparency}{numeric value used to set transparency of intervals.
0 means fully transparent, 1 means opaque. Default to \code{0.25}}

\item{intervals}{a vector of \code{numeric} values indicating which central
prediction interval levels to plot. \code{NULL} means no interval levels.
If not provided, it will default to \code{c(.5, .8, .95)}.
When plotting 6 models or more, the plot will be reduced to show \code{.95}
interval only. Value possibles: \verb{0.5, 0.8, 0.9, 0.95}}

\item{top_layer}{character vector, where the first element indicates the top
layer of the resulting plot. Possible options are \code{"forecast"} (default)
and \code{"truth"}}

\item{title}{a \code{character} string, if not NULL, will be added as title to the
plot}

\item{ens_color}{a \code{character} string of a color name, if not NULL, will be
use as color for the model name associated with the parameter \code{ens_name}
(both parameter need to be provided)}

\item{ens_name}{a \code{character} string of a model name, if not NULL, will be
use to change the color for the model name, associated with the parameter
\code{ens_color}(both parameter need to be provided)}
}
\description{
Create a simple Plotly time-series plot for model projection outputs.
}
