% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_step_ahead_model_output.R
\name{plot_step_ahead_model_output}
\alias{plot_step_ahead_model_output}
\title{Basic Plot for model outputs}
\usage{
plot_step_ahead_model_output(
  model_output_data,
  truth_data,
  use_median_as_point = FALSE,
  plot = TRUE,
  plot_truth = TRUE,
  x_col_name = "target_date",
  x_truth_col_name = "time_idx",
  show_legend = TRUE,
  facet = NULL,
  facet_scales = "fixed",
  facet_nrow = NULL,
  facet_ncol = NULL,
  facet_title = "top left",
  interactive = TRUE,
  fill_by = "model_id",
  pal_color = "Set2",
  one_color = "blue",
  fill_transparency = 0.25,
  intervals = c(0.5, 0.8, 0.95),
  top_layer = "model_output",
  title = NULL,
  ens_color = NULL,
  ens_name = NULL
)
}
\arguments{
\item{model_output_data}{a \code{model_out_tbl} object, containing all the required
columns, a column containing date information (\code{x_col_name} parameter) and
a column \code{value}.}

\item{truth_data}{a \code{data.frame} object containing the ground truth data,
with a column containing date information (\code{x_truth_col_name} parameter) and
a column \code{value}. Ignored, if \code{plot_truth = FALSE}.}

\item{use_median_as_point}{a \code{Boolean} for using median quantile as point
in plot. Default to FALSE. If TRUE, will select first any \code{median}
output type value and if no \code{median} value included in \code{model_output_data};
will select \code{quantile = 0.5} output type value.}

\item{plot}{a \code{boolean} for showing the plot. Default to TRUE.}

\item{plot_truth}{a \code{boolean} for showing the truth data in the plot.
Default to TRUE. Data used in the plot comes from the parameter \code{truth_data}}

\item{x_col_name}{column name containing the date information for \code{all_plot}
and \code{all_ens} data frames, value will be map to the x-axis of the plot.
By default, "target_date".}

\item{x_truth_col_name}{column name containing the date information for
\code{truth_data} data frame, value will be map to the x-axis of the plot.
By default, "time_idx".}

\item{show_legend}{a \code{boolean} for showing the legend in the plot.
Default to TRUE.}

\item{facet}{a unique value corresponding as a task_id variable name
(interpretable as facet option for ggplot)}

\item{facet_scales}{argument for scales as in \link[ggplot2:facet_wrap]{ggplot2::facet_wrap} or
equivalent to \code{shareX}, \code{shareY} in \link[plotly:subplot]{plotly::subplot}. Default to "fixed"
(x and y axes are shared).}

\item{facet_nrow}{a numeric, number of rows in the layout.}

\item{facet_ncol}{a numeric, number of columns in the layout
(ignored in \link[plotly:subplot]{plotly::subplot})}

\item{facet_title}{a \code{string}, position of each subplot tile (value
associated with the \code{facet} parameter). "top right", "top left" (default),
"bottom right", "bottom left" are the possible values, \code{NULL} to remove the
title. For interactive plot only.}

\item{interactive}{a \code{boolean} to output an "interactive" version of the
plot (using Plotly) or a "static" plot (using ggplot2). By default, \code{TRUE}
(interactive plot)}

\item{fill_by}{name of a column for specifying colors and legend in plot.
The \code{pal_color} parameter can be use to change the palette.
Default to \code{model_id}.}

\item{pal_color}{a \code{character} string for specifying the palette color in the
plot. Please refer to \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::display.brewer.all()}}. If \code{NULL},
only \code{one_color} parameter will be used for all models. Default to \code{"Set2"}}

\item{one_color}{a \code{character} string for specifying the color in the
plot if \code{pal_color} is set to \code{NULL}. Please refer
to \code{\link[=colors]{colors()}} for accepted color names. Default to \code{"blue"}}

\item{fill_transparency}{numeric value used to set transparency of intervals.
0 means fully transparent, 1 means opaque. Default to \code{0.25}}

\item{intervals}{a vector of \code{numeric} values indicating which central
prediction interval levels to plot. \code{NULL} means no interval levels.
If not provided, it will default to \code{c(.5, .8, .95)}.
When plotting 6 models or more, the plot will be reduced to show \code{.95}
interval only. Value possibles: \verb{0.5, 0.8, 0.9, 0.95}}

\item{top_layer}{character vector, where the first element indicates the top
layer of the resulting plot. Possible options are \code{"model_output"} (default)
and \code{"truth"}}

\item{title}{a \code{character} string, if not NULL, will be added as title to the
plot}

\item{ens_color}{a \code{character} string of a color name, if not NULL, will be
use as color for the model name associated with the parameter \code{ens_name}
(both parameter need to be provided)}

\item{ens_name}{a \code{character} string of a model name, if not NULL, will be
use to change the color for the model name, associated with the parameter
\code{ens_color}(both parameter need to be provided)}
}
\description{
Create a simple Plotly time-series plot for model projection outputs.
}
